improved files and added some command line arguments

+++ iput_iget_tests/iput_iget.sh
@@ -1,35 +1,47 @@
 #!/bin/bash
 
 RESOURCE=replResc
-N=16 # num files
 
+re='^[0-9]+$'
+if ! [[ $2 =~ $re ]] ; then
+    echo "error: Not a number" >&2; exit 1
+fi
+
+if ! [ "$1" ==  "iput" ] || [ "$1" == "iget" ] ;then
+    echo "error: Not 'iput' or 'iget'" 
+    exit 1
+fi
+
+if [ "$2" -lt "10" ]; then
+    N=0$2
+else 
+    N=$2
+fi
 mkdir -p /tmp/benchmarks.dir
-imkdir -p benchmarks.coll
+
 
 cd /tmp/random_files.dir
 
 rm /tmp/benchmarks.dir/iput-all /tmp/benchmarks.dir/iget-all
 for i in `seq -w $N`
 do
-    echo iput -R ${RESOURCE} -K -f $PWD/random_file.128M.$i  benchmarks.coll/ >> /tmp/benchmarks.dir/iput-all
-    echo iget -R ${RESOURCE} -K -f benchmarks.coll/random_file.128M.$i >> /tmp/benchmarks.dir/iget-all
+    echo iput -R ${RESOURCE} -K -f $PWD/random_file_128M_$i   >> /tmp/benchmarks.dir/iput-all
+    echo iget -R ${RESOURCE} -K -f random_file_128M_$i >> /tmp/benchmarks.dir/iget-all
 done
 
 cd /tmp/benchmarks.dir
 
-for mode in "iput" "iget"
-do
+
+
   for i in `seq -w $N`
   do
-      echo $i $mode jobs:
+      echo $i $1 jobs:
       date
-      cat $mode"-all" | head -$i | parallel --jobs $i --joblog $mode.$i.log
+      cat $1"-all" | head -$i | parallel --jobs $i --joblog $1.$i.log
       date
-      md5sum /tmp/random_files.dir/random_file.128M.* | head -$i > /tmp/benchmark.md5
-      echo $( cat /tmp/random_files.md5 | head -$i ) 
-      echo $( cat /tmp/benchmark.md5 | head -$i )
+      md5sum /tmp/random_files.dir/random_file_128M_* | head -$i > /tmp/benchmark.md5
       cat /tmp/random_files.md5 | head -$i > /tmp/sorted_random_files.md5
       echo Checksum errors: $(comm -2 -3 /tmp/sorted_random_files.md5   /tmp/benchmark.md5 | wc -l )    
       sleep 5
-  done
+
 done 2>&1 | tee benchmark.log
